Moved informal test data into separate file, commented quell-server code, and replaced dummyCache commands with Redis. Installed redis library.
