+++ quell-server/package.json
@@ -9,6 +9,7 @@
   "author": "Quell",
   "license": "ISC",
   "dependencies": {
-    "graphql": "^15.3.0"
+    "graphql": "^15.3.0",
+    "redis": "^3.0.2"
   }
 }
